Extract tests for isIntlCollator to own file
