+++ lib/referee.test.js
@@ -4,42 +4,6 @@ var referee = require("../lib/referee");
 var testHelper = require("./test-helper");
 var sinon = require("sinon");
 
-testHelper.assertionTests("assert", "isIntlCollator", function(
-    pass,
-    fail,
-    msg,
-    error
-) {
-    function captureArgs() {
-        return arguments;
-    }
-
-    pass("for Intl.Collator", new Intl.Collator());
-    fail("for string", "apple pie");
-    fail("for array", []);
-    fail("for object", {});
-    fail("for arguments", captureArgs());
-    msg(
-        "fail with descriptive message",
-        "[assert.isIntlCollator] Expected {  } to be an Intl.Collator",
-        {}
-    );
-    msg(
-        "fail with custom message",
-        "[assert.isIntlCollator] Nope: Expected {  } to be an Intl.Collator",
-        {},
-        "Nope"
-    );
-    error(
-        "for object",
-        {
-            code: "ERR_ASSERTION",
-            operator: "assert.isIntlCollator"
-        },
-        {}
-    );
-});
-
 testHelper.assertionTests("assert", "isIntlDateTimeFormat", function(
     pass,
     fail,
